python/requirements.txt
# PDF 처리 및 OCR
PyMuPDF
pdf2image
Pillow
google-cloud-vision

# 데이터 처리
pandas
tabulate
tqdm

# LangChain 및 관련 기능
langchain
langchain-community
langchain-core
langchain-chroma
chromadb

# 임베딩 모델 (HuggingFace)
sentence-transformers
transformers
huggingface-hub

# OpenAI (LLM 사용)
openai

# FastAPI + Middleware
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
aiofiles
httpx

# 환경 변수 로드
python-dotenv

# 스케줄러
apscheduler

# OpenTelemetry + Chroma 의존성 충돌 해결용
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
protobuf